move markdown report generator to bundlemon utils
