+++ packages/bundlemon-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "bundlemon-utils",
-  "version": "0.3.2",
+  "version": "0.3.3",
   "description": "",
   "keywords": [],
   "author": "Liron Er",
