+++ packages/bundlemon-utils/src/index.ts
@@ -4,3 +4,4 @@ export * from './consts';
 export { generateDiffReport } from './diffReport';
 export * from './types';
 export * as textUtils from './textUtils';
+export { generateReportMarkdown } from './markdown/markdownReportGenerator';
