+++ packages/bundlemon/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "axios": "^0.21.1",
     "brotli-size": "^4.0.0",
-    "bundlemon-utils": "^0.3.2",
+    "bundlemon-utils": "^0.3.3",
     "bytes": "^3.1.0",
     "chalk": "^4.1.1",
     "commander": "^8.0.0",
